+++ resources/js/login.js
@@ -1,3 +1,5 @@
+import LocalizationSelect from "./components/tools/LocalizationSelect";
+
 window.$ = require('jquery');
 window.Vue = require('vue');
 window.Buefy = require('buefy');
@@ -6,6 +8,7 @@ window.Buefy = require('buefy');
 const Content = new Vue({
     el: '#sectionContent',
     components: {
+        LocalizationSelect
     },
     data: {
     },